pyproject.toml
[tool.poetry]
name = "s3-transfer-service"
version = "0.1.0"
description = ""
authors = ["Mattias Poppe <mattias.poppe@meemoo.be>"]

[tool.poetry.dependencies]
python = "^3.6"
pika = "^1.2.0"
paramiko = "^2.7.2"
viaa-chassis = {version = "^0.2.0rc1", allow-prereleases = true}
requests = "^2.25.1"
retry = "^0.9.2"
pulsar-client = "^2.8.1"
hvac = "^0.11.2"
meemoo-cloudevents = "0.1.0rc1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
flake8 = "^3.8.4"
black = "^20.8b1"

[[tool.poetry.source]]
name = "meemoo"
url = "http://do-prd-mvn-01.do.viaa.be:8081/repository/pypi-all/simple"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
